# mcp-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProTour Data API",
    "version": "0.1.0",
    "description": "REST API for querying Pokémon TCG ProTour tournament data including matches, deck lists, and statistics.",
    "contact": {
      "name": "API Support"
    },
    "x-privacy-policy-url": "https://github.com/sp3c1/protour-data-viz#privacy"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/matches": {
      "get": {
        "summary": "Query tournament matches",
        "description": "Query matches by round, player name, or archetype",
        "operationId": "queryMatches",
        "parameters": [
          {
            "name": "round",
            "in": "query",
            "description": "Round number (1-20)",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20
            }
          },
          {
            "name": "player",
            "in": "query",
            "description": "Player name (partial match supported)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "archetype",
            "in": "query",
            "description": "Deck archetype (partial match supported)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/decks": {
      "get": {
        "summary": "Get deck lists with full card details",
        "description": "Retrieve complete deck lists including main deck, sideboard, and all card details. Filter by player name or archetype to find specific decks.",
        "operationId": "queryDecks",
        "parameters": [
          {
            "name": "player",
            "in": "query",
            "description": "Player name (partial match supported)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "archetype",
            "in": "query",
            "description": "Deck archetype (partial match supported)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecksResponse"
                }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Get archetype win rates and matchup statistics",
        "description": "Retrieve detailed statistics including win/loss records, win rates, and head-to-head matchup data for any archetype. Essential for meta analysis.",
        "operationId": "queryStats",
        "parameters": [
          {
            "name": "archetype",
            "in": "query",
            "description": "Specific archetype name (returns all if omitted)",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Archetype not found"
          }
        }
      }
    },
    "/players/{player}/deck": {
      "get": {
        "summary": "Get specific player's complete deck list and tournament results",
        "description": "Retrieve a player's full deck list (main deck and sideboard), their archetype, and all match results. Perfect for analyzing individual player performance.",
        "operationId": "queryPlayerDeck",
        "parameters": [
          {
            "name": "player",
            "in": "path",
            "required": true,
            "description": "Player name (exact or partial match)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDeckResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found"
          }
        }
      }
    },
    "/players/{player}/stats": {
      "get": {
        "summary": "Get individual player's performance statistics",
        "description": "Retrieve detailed performance stats for a specific player including their W-L-D record, win rate, matchup breakdown, and deck list. Shows individual tournament performance.",
        "operationId": "queryPlayerStats",
        "parameters": [
          {
            "name": "player",
            "in": "path",
            "required": true,
            "description": "Player name (exact or partial match)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerStatsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Player not found"
          }
        }
      }
    },
    "/archetypes": {
      "get": {
        "summary": "List all archetypes",
        "description": "Get list of all deck archetypes with statistics",
        "operationId": "listArchetypes",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchetypesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/tournament": {
      "get": {
        "summary": "Get tournament information",
        "description": "Get tournament metadata and summary statistics",
        "operationId": "getTournamentInfo",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentInfoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card}": {
      "get": {
        "summary": "Query decks by card name",
        "description": "Find all decks containing a specific card and get their performance statistics. Returns archetype breakdown and win rates for decks playing this card.",
        "operationId": "queryDecksByCard",
        "parameters": [
          {
            "name": "card",
            "in": "path",
            "required": true,
            "description": "Card name (partial match supported, e.g., 'Badgermole Cub')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of deck lists to return",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDecksResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid card name"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MatchesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "count": {
            "type": "integer",
            "example": 10
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "DecksResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "count": {
            "type": "integer",
            "example": 5
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          }
        }
      },
      "PlayerDeckResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          }
        }
      },
      "PlayerStatsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "playerName": {"type": "string"},
              "archetype": {"type": "string"},
              "wins": {"type": "integer"},
              "losses": {"type": "integer"},
              "draws": {"type": "integer"},
              "winRate": {"type": "number"},
              "totalMatches": {"type": "integer"},
              "record": {"type": "string", "example": "5-2"},
              "matchups": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "opponent": {"type": "string"},
                    "wins": {"type": "integer"},
                    "losses": {"type": "integer"},
                    "draws": {"type": "integer"},
                    "percentage": {"type": "number"}
                  }
                }
              },
              "deckList": {"type": "object"}
            }
          }
        }
      },
      "ArchetypesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "count": {
            "type": "integer",
            "example": 43
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "TournamentInfoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object"
          }
        }
      },
      "CardDecksResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "card": {
                "type": "string",
                "example": "Badgermole Cub"
              },
              "totalDecks": {
                "type": "integer",
                "example": 15
              },
              "overallStats": {
                "type": "object",
                "properties": {
                  "wins": {"type": "integer"},
                  "losses": {"type": "integer"},
                  "draws": {"type": "integer"},
                  "winRate": {"type": "number"}
                }
              },
              "archetypeBreakdown": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "decks": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Invalid request"
          }
        }
      }
    }
  }
}